// types/job.ts
export interface Job {
  jobId: string;
  title: string;
  company: string;
  location: string;
  job_link: string;
  employment_type: string;
  experience: string;
  source: string;
  country: string;
  postedDateTime: string; 
  companyImageUrl: string;
  company_url: string;
  seniority_level: string;
  min_exp: number;
  max_exp: number;
  companytype: string;
}